requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
pypdf>=3.17.0
pyyaml>=6.0.1
openai>=1.6.0
tiktoken>=0.5.2
